+++ classes/class.ilSrSearchPDBlockUIHookGUI.php
@@ -1,9 +1,11 @@
 <?php
 
 use srag\DIC\SrSearchPDBlock\DICTrait;
-use srag\DIC\SrSearchPDBlock\Exception\DICException;
-use srag\Plugins\SrSearchPDBlock\Block\SearchBlock53;
-use srag\Plugins\SrSearchPDBlock\Block\SearchBlock54;
+use srag\Plugins\SrSearchPDBlock\Block\CurrentPageSearch\CurrentPageSearchBlock53;
+use srag\Plugins\SrSearchPDBlock\Block\CurrentPageSearch\CurrentPageSearchBlock54;
+use srag\Plugins\SrSearchPDBlock\Block\GlobalSearch\GlobalSearchBlock53;
+use srag\Plugins\SrSearchPDBlock\Block\GlobalSearch\GlobalSearchBlock54;
+use srag\Plugins\SrSearchPDBlock\Config\Config;
 
 /**
  * Class ilSrSearchPDBlockUIHookGUI
@@ -15,16 +17,9 @@ class ilSrSearchPDBlockUIHookGUI extends ilUIHookPluginGUI
 
     use DICTrait;
     const PLUGIN_CLASS_NAME = ilSrSearchPDBlockPlugin::class;
-    const PERSONAL_DESKTOP_INIT = "personal_desktop";
+    const COMPONENT_CONTAINER = "Services/Container";
     const COMPONENT_PERSONAL_DESKTOP = "Services/PersonalDesktop";
     const PART_CENTER_RIGHT = "right_column";
-    /**
-     * @var bool[]
-     */
-    protected static $load
-        = [
-            self::PERSONAL_DESKTOP_INIT => false
-        ];
 
 
     /**
@@ -33,9 +28,6 @@ class ilSrSearchPDBlockUIHookGUI extends ilUIHookPluginGUI
      * @param array  $a_par
      *
      * @return array
-     *
-     * @throws DICException
-     * @throws ilTemplateException
      */
     public function getHTML(/*string*/
         $a_comp, /*string*/
@@ -43,19 +35,42 @@ class ilSrSearchPDBlockUIHookGUI extends ilUIHookPluginGUI
         $a_par = []
     ) : array {
 
-        if (!self::$load[self::PERSONAL_DESKTOP_INIT]) {
+        if ($a_comp === self::COMPONENT_PERSONAL_DESKTOP && $a_part === self::PART_CENTER_RIGHT) {
+            return [
+                "mode" => self::PREPEND,
+                "html" => $this->getBlocks(Config::KEY_SHOW_GLOBAL_SEARCH_PERSONAL_DESKTOP, Config::KEY_SHOW_CURRENT_PAGE_SEARCH_PERSONAL_DESKTOP)
+            ];
+        }
 
-            if ($a_comp === self::COMPONENT_PERSONAL_DESKTOP && $a_part === self::PART_CENTER_RIGHT) {
-
-                self::$load[self::PERSONAL_DESKTOP_INIT] = true;
-
-                return [
-                    "mode" => self::PREPEND,
-                    "html" => self::output()->getHTML(self::version()->is54() ? new SearchBlock54() : new SearchBlock53())
-                ];
-            }
+        if ($a_comp === self::COMPONENT_CONTAINER && $a_part === self::PART_CENTER_RIGHT) {
+            return [
+                "mode" => self::PREPEND,
+                "html" => $this->getBlocks(Config::KEY_SHOW_GLOBAL_SEARCH_CONTAINER_OBJECTS, Config::KEY_SHOW_CURRENT_PAGE_SEARCH_CONTAINER_OBJECTS)
+            ];
         }
 
         return parent::getHTML($a_comp, $a_part, $a_par);
+    }
+
+
+    /**
+     * @param string $key_config_global_search
+     * @param string $key_config_current_page_search
+     *
+     * @return string
+     */
+    protected function getBlocks(string $key_config_global_search, string $key_config_current_page_search) : string
+    {
+        $blocks = [];
+
+        if (Config::getField($key_config_global_search)) {
+            $blocks[] = self::version()->is54() ? new GlobalSearchBlock54() : new GlobalSearchBlock53();
+        }
+
+        if (Config::getField($key_config_current_page_search)) {
+            $blocks[] = self::version()->is54() ? new CurrentPageSearchBlock54() : new CurrentPageSearchBlock53();
+        }
+
+        return self::output()->getHTML($blocks);
     }
 }