+++ src/Block/BaseSearchBlock.php
@@ -3,9 +3,8 @@
 namespace srag\Plugins\SrSearchPDBlock\Block;
 
 use ilBlockGUI;
-use ilSearchController;
 use ilSrSearchPDBlockPlugin;
-use ilTemplateException;
+use ilTemplate;
 use srag\DIC\SrSearchPDBlock\DICTrait;
 use srag\DIC\SrSearchPDBlock\Exception\DICException;
 
@@ -21,6 +20,12 @@ abstract class BaseSearchBlock extends ilBlockGUI
 
     use DICTrait;
     const PLUGIN_CLASS_NAME = ilSrSearchPDBlockPlugin::class;
+    /**
+     * @var string
+     *
+     * @abstract
+     */
+    const LANG_MODULE = "";
 
 
     /**
@@ -37,35 +42,38 @@ abstract class BaseSearchBlock extends ilBlockGUI
 
 
     /**
-     * @throws DICException
+     *
      */
     protected function initBlock()/*: void*/
     {
         self::dic()->mainTemplate()->addCss(self::plugin()->directory() . "/css/srsearchpdblock.css");
 
-        $this->setTitle(self::dic()->language()->txt("search"));
+        $this->setTitle(self::plugin()->translate("title", static::LANG_MODULE));
     }
 
 
     /**
-     * @throws DICException
-     * @throws ilTemplateException
+     *
      */
     public function fillDataSection()/*: void*/
     {
-        $tpl = self::plugin()->template("search_input.html");
-
-        $tpl->setVariable("TXT_PLACEHOLDER", self::dic()->language()->txt("search") . " ...");
-        $tpl->setVariable("TXT_GO", self::dic()->language()->txt("go"));
-
-        //Services/Search/classes/class.ilMainMenuSearchGUI.php::getHTML
-        $search_action = "ilias.php?baseClass=" . ilSearchController::class . "&cmd=post&rtoken=" . self::dic()->ctrl()->getRequestToken()
-            . "&fallbackCmd=remoteSearch";
-        $tpl->setVariable("ACTION", $search_action);
-
-        $tpl->setVariable("BUTTON", self::output()->getHTML(self::dic()->ui()->factory()->button()->standard(self::dic()->language()
-            ->txt("search"), "")));
-
-        $this->setDataSection(self::output()->getHTML($tpl));
+        $this->setDataSection(self::output()->getHTML($this->fillTemplate($this->getTemplate())));
     }
+
+
+    /**
+     * @return ilTemplate
+     */
+    protected function getTemplate() : ilTemplate
+    {
+        return self::plugin()->template(static::LANG_MODULE . ".html");
+    }
+
+
+    /**
+     * @param ilTemplate $tpl
+     *
+     * @return ilTemplate
+     */
+    protected abstract function fillTemplate(ilTemplate $tpl) : ilTemplate;
 }